11-3 - рефакторинг
